requirements.txt
bcrypt==2.0.0
cffi==1.3.0
Django==1.9.5
djangorestframework==3.3.2
funcsigs==0.4
mock==1.3.0
passlib==1.6.5
pbr==1.8.1
psycopg2==2.6.1
pycparser==2.14
six==1.10.0
wheel==0.24.0
mysql-python=1.2.5